file_export: separated codecs and reorganized into codecs/ directory
* updated configure.ac and Makefile.am to handle new structure
* split up image codecs into their own files and updated screenshot code
