--- src/audio_codec_mulaw.h
@@ -1,10 +1,0 @@
-#ifndef AUDIO_CODEC_MULAW_H_
-#define AUDIO_CODEC_MULAW_H_
-
-#include "atari.h"
-#include "file_export.h"
-
-extern AUDIO_CODEC_t Audio_Codec_MULAW;
-extern AUDIO_CODEC_t Audio_Codec_PCM_MULAW;
-
-#endif /* AUDIO_CODEC_MULAW_H_ */
