--- src/audio_codec_pcm.h
@@ -1,9 +1,0 @@
-#ifndef AUDIO_CODEC_PCM_H_
-#define AUDIO_CODEC_PCM_H_
-
-#include "atari.h"
-#include "file_export.h"
-
-extern AUDIO_CODEC_t Audio_Codec_PCM;
-
-#endif /* AUDIO_CODEC_PCM_H_ */
