+++ src/codecs/audio_mulaw.h
@@ -1,0 +1,10 @@
+#ifndef CODECS_AUDIO_MULAW_H_
+#define CODECS_AUDIO_MULAW_H_
+
+#include "atari.h"
+#include "codecs/audio.h"
+
+extern AUDIO_CODEC_t Audio_Codec_MULAW;
+extern AUDIO_CODEC_t Audio_Codec_PCM_MULAW;
+
+#endif /* CODECS_AUDIO_MULAW_H_ */
