+++ src/codecs/audio_pcm.h
@@ -1,0 +1,9 @@
+#ifndef CODECS_AUDIO_PCM_H_
+#define CODECS_AUDIO_PCM_H_
+
+#include "atari.h"
+#include "codecs/audio.h"
+
+extern AUDIO_CODEC_t Audio_Codec_PCM;
+
+#endif /* CODECS_AUDIO_PCM_H_ */
