+++ src/codecs/image_pcx.h
@@ -1,0 +1,9 @@
+#ifndef CODECS_IMAGE_PCX_H_
+#define CODECS_IMAGE_PCX_H_
+
+#include "atari.h"
+#include "codecs/image.h"
+
+extern IMAGE_CODEC_t Image_Codec_PCX;
+
+#endif /* CODECS_IMAGE_PCX_H_ */
