+++ src/codecs/image_png.h
@@ -1,0 +1,9 @@
+#ifndef CODECS_IMAGE_PNG_H_
+#define CODECS_IMAGE_PNG_H_
+
+#include "atari.h"
+#include "codecs/image.h"
+
+extern IMAGE_CODEC_t Image_Codec_PNG;
+
+#endif /* CODECS_IMAGE_PNG_H_ */
