+++ src/codecs/video_mpng.h
@@ -1,0 +1,10 @@
+#ifndef CODECS_VIDEO_MPNG_H_
+#define CODECS_VIDEO_MPNG_H_
+
+#include "atari.h"
+#include "codecs/video.h"
+
+extern VIDEO_CODEC_t Video_Codec_MPNG;
+
+#endif /* CODECS_VIDEO_MPNG_H_ */
+
