+++ src/codecs/video_mrle.c
@@ -1,5 +1,5 @@
 /*
- * video_codec_mrle.c - Video codec for Microsoft Run-Length Encoding format
+ * video_mrle.c - Video codec for Microsoft Run-Length Encoding format
  *
  * Copyright (C) 2021 Rob McMullen
  * Copyright (C) 1998-2021 Atari800 development team (see DOC/CREDITS)
@@ -25,8 +25,8 @@
 #include <string.h>
 #include <stdlib.h>
 #include <math.h>
-#include "file_export.h"
-#include "video_codec_mrle.h"
+#include "codecs/video.h"
+#include "codecs/video_mrle.h"
 #include "screen.h"
 #include "util.h"
