+++ src/codecs/video_mrle.h
@@ -1,0 +1,11 @@
+#ifndef CODECS_VIDEO_MRLE_H_
+#define CODECS_VIDEO_MRLE_H_
+
+#include "atari.h"
+#include "codecs/video.h"
+
+extern VIDEO_CODEC_t Video_Codec_MRLE;
+extern VIDEO_CODEC_t Video_Codec_MSRLE;
+
+#endif /* CODECS_VIDEO_MRLE_H_ */
+
