+++ src/codecs/video_zmbv.h
@@ -1,0 +1,10 @@
+#ifndef CODECS_VIDEO_ZMBV_H_
+#define CODECS_VIDEO_ZMBV_H_
+
+#include "atari.h"
+#include "codecs/video.h"
+
+extern VIDEO_CODEC_t Video_Codec_ZMBV;
+
+#endif /* CODECS_VIDEO_ZMBV_H_ */
+
