--- src/video_codec_mpng.h
@@ -1,10 +1,0 @@
-#ifndef VIDEO_CODEC_MPNG_H_
-#define VIDEO_CODEC_MPNG_H_
-
-#include "atari.h"
-#include "file_export.h"
-
-extern VIDEO_CODEC_t Video_Codec_MPNG;
-
-#endif /* VIDEO_CODEC_MPNG_H_ */
-
