--- src/video_codec_mrle.h
@@ -1,11 +1,0 @@
-#ifndef VIDEO_CODEC_MRLE_H_
-#define VIDEO_CODEC_MRLE_H_
-
-#include "atari.h"
-#include "file_export.h"
-
-extern VIDEO_CODEC_t Video_Codec_MRLE;
-extern VIDEO_CODEC_t Video_Codec_MSRLE;
-
-#endif /* VIDEO_CODEC_MRLE_H_ */
-
