--- src/video_codec_zmbv.h
@@ -1,10 +1,0 @@
-#ifndef VIDEO_CODEC_ZMBV_H_
-#define VIDEO_CODEC_ZMBV_H_
-
-#include "atari.h"
-#include "file_export.h"
-
-extern VIDEO_CODEC_t Video_Codec_ZMBV;
-
-#endif /* VIDEO_CODEC_ZMBV_H_ */
-
